feat!: consistent naming convention
