+++ collectivo/collectivo/server/plugins/registerExtension.ts
@@ -1,12 +1,20 @@
 import pkg from "../../package.json";
 
-import m001_extensions from "../schemas/001_extensions";
-import m002_settings from "../schemas/002_settings";
-import m003_tags from "../schemas/003_tags";
-import m004_tiles from "../schemas/004_tiles";
-import m005_messages from "../schemas/005_messages";
+import settings_01 from "../schemas/settings_01";
+import extensions_01 from "../schemas/extensions_01";
+import tags_01 from "../schemas/tags_01";
+import tiles_01 from "../schemas/tiles_01";
+import messages_01 from "../schemas/messages_01";
 
 import examples from "../examples/examples";
+
+const schema_0_0_1 = combineSchemas("collectivo", "0.0.1", [
+  settings_01,
+  extensions_01,
+  tags_01,
+  tiles_01,
+  messages_01,
+]);
 
 // Register extension on startup
 export default defineNitroPlugin(() => {
@@ -15,15 +23,7 @@ export default defineNitroPlugin(() => {
   registerExtension({
     name: "collectivo",
     version: pkg.version,
-    schemas: [
-      combineSchemas("collectivo", "0.0.1", [
-        m001_extensions,
-        m002_settings,
-        m003_tags,
-        m004_tiles,
-        m005_messages,
-      ]),
-    ],
+    schemas: [schema_0_0_1],
     examples: examples,
   });
 });
